backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Bcrypt==1.0.1
opensearch-py==2.4.0
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
aiohttp
